requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Brotli==1.0.9
cffi==1.15.1
charset-normalizer==2.1.1
Django==4.1.3
dnspython==2.2.1
fake-useragent==0.1.11
frozenlist==1.3.1
greenlet==2.0.0
idna==3.4
loguru==0.6.0
multidict==6.0.2
peewee==3.15.3
psycopg2-binary==2.9.5
pycares==4.2.2
pycparser==2.21
pydantic==1.10.2
pymongo==4.3.2
python-dotenv==0.21.0
SQLAlchemy==2.0.0b2
sqlparse==0.4.3
typing_extensions==4.4.0
yarl==1.8.1
